First pass at decorating associations
This is a little bit hacky, but its up and running to start.
Still need to implement the ability to decorate associations
that return an array or relation, as currently its assumed
to just be a single object and handled accordingly.

I also suspect this will need to be revamped to use the
"eigenclass" (class << self) to ensure instance variables
and their getters aren't defined on a class level and
instead as their singletons.