Remove decorated associations
In an attempt to keep this gem as simple as possible,
I've opted to remove the automatic decoration of associations.
I believe this will keep things more explicit and remove a
bit of the magic that was occurring.

Decoration of associations can still occur, only they will need
to be explicity called. Something along the lines of:

    def account
      decorate(@delegated_object.account)
    end

***

Related #3, #10